move date format function to a ts module
